Updated Transport to a protocol from class.  Added new transport to use NSURLSession. Updated NSURLSession to be more block based and event driven removing CPU burning loops. Added.  Added data support on Handshake.  Support reconnection on transport errors in case switching from WiFi to Wireless.  Example reachability support added for automatic reconnection when losing and gaining network connectivity.  Updated CometClient.State changes to allow KVO.

+++ DDComet/DDCometSubscription.m
@@ -93,7 +93,7 @@
 
 -(NSString*)description
 {
-    return [NSString stringWithFormat:@"%@-%d", self.channel, self.hash];
+    return [NSString stringWithFormat:@"%@-%lu", self.channel, (unsigned long)self.hash];
 }
 
 @end
